fixed lint too many functions error for TwitchDao
